rename: 인수 테스트 파일은 app/admin에 따라 파일 이동
