+++ be/promotion/src/test/java/woowa/promotion/acceptance/admin/CouponGroupAcceptanceTest.java
@@ -1,4 +1,4 @@
-package woowa.promotion.acceptance;
+package woowa.promotion.acceptance.admin;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
