+++ be/promotion/src/test/java/woowa/promotion/acceptance/admin/PromotionAcceptanceTest.java
@@ -1,6 +1,19 @@
-package woowa.promotion.acceptance;
+package woowa.promotion.acceptance.admin;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static woowa.promotion.fixture.CouponGroupFixture.A_쿠폰그룹;
+import static woowa.promotion.fixture.CouponGroupFixture.B_쿠폰그룹;
+import static woowa.promotion.fixture.FixtureFactory.createCouponGroup;
+import static woowa.promotion.fixture.FixtureFactory.createPromotion;
+import static woowa.promotion.fixture.FixtureFactory.createPromotionOptionRequest;
+import static woowa.promotion.fixture.PromotionFixture.A_프로모션;
+import static woowa.promotion.fixture.PromotionFixture.B_프로모션;
+import static woowa.promotion.fixture.PromotionOptionFixture.A_프로모션_옵션;
+import static woowa.promotion.fixture.PromotionOptionFixture.B_프로모션_옵션;
 
 import io.restassured.RestAssured;
+import java.util.List;
+import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
@@ -8,18 +21,6 @@ import org.springframework.http.MediaType;
 import woowa.promotion.admin.admin.domain.Admin;
 import woowa.promotion.fixture.FixtureFactory;
 import woowa.promotion.util.AcceptanceTest;
-
-import java.util.List;
-import java.util.Map;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static woowa.promotion.fixture.CouponGroupFixture.A_쿠폰그룹;
-import static woowa.promotion.fixture.CouponGroupFixture.B_쿠폰그룹;
-import static woowa.promotion.fixture.FixtureFactory.*;
-import static woowa.promotion.fixture.PromotionFixture.A_프로모션;
-import static woowa.promotion.fixture.PromotionFixture.B_프로모션;
-import static woowa.promotion.fixture.PromotionOptionFixture.A_프로모션_옵션;
-import static woowa.promotion.fixture.PromotionOptionFixture.B_프로모션_옵션;
 
 @DisplayName("[인수테스트] 관리자 - 프로모션")
 public class PromotionAcceptanceTest extends AcceptanceTest {
